requirements.txt
gradio>=3.50
torch>=2.0
torchvision
transformers>=4.36
pillow
numpy
scikit-learn
matplotlib
timm
pandas